chore(a11y-testing-spike): move implementation to 2nd gen, move docs
