+++ 1st-gen/package.json
@@ -62,11 +62,11 @@
         "storybook:quick": "run-p build:watch storybook:run",
         "storybook:run": "web-dev-server --config wds-storybook.config.js",
         "test": "yarn test:focus unit",
-        "test:a11y": "playwright test --config=playwright.a11y.config.ts",
-        "test:a11y:ui": "playwright test --config=playwright.a11y.config.ts --ui",
-        "test:a11y:1st": "playwright test --config=playwright.a11y.config.ts --project=1st-gen",
-        "test:a11y:2nd": "playwright test --config=playwright.a11y.config.ts --project=2nd-gen",
-        "test:a11y:report": "playwright show-report test/playwright-a11y/report",
+        "test:a11y": "playwright test --config=../2nd-gen/playwright.a11y.config.ts",
+        "test:a11y:1st": "playwright test --config=../2nd-gen/playwright.a11y.config.ts --project=1st-gen",
+        "test:a11y:2nd": "playwright test --config=../2nd-gen/playwright.a11y.config.ts --project=2nd-gen",
+        "test:a11y:report": "playwright show-report ../2nd-gen/test/playwright-a11y/report",
+        "test:a11y:ui": "playwright test --config=../2nd-gen/playwright.a11y.config.ts --ui",
         "test:bench": "yarn build:tests && node test/benchmark/cli.js",
         "test:changed": "node ./scripts/test-changes.js",
         "test:ci": "yarn test:start",
