+++ 1st-gen/tsconfig-all.json
@@ -1,21 +1,21 @@
 {
     "extends": "./tsconfig.json",
     "compilerOptions": {
-        "rootDir": "./",
+        "emitDeclarationOnly": false,
         "noEmit": true,
-        "emitDeclarationOnly": false
+        "rootDir": "./"
     },
-    "include": [
-        "packages/**/*.ts",
-        "tools/**/*.ts",
-        "projects/story-decorator/**/*.ts",
-        "projects/vrt-compare/**/*.ts"
-    ],
     "exclude": [
         "packages/*/node_modules/**/*.ts",
         "tools/*/node_modules/**/*.ts",
         "packages/**/test/*.a11y.spec.ts",
         "test/**/*.ts"
+    ],
+    "include": [
+        "packages/**/*.ts",
+        "tools/**/*.ts",
+        "projects/story-decorator/**/*.ts",
+        "projects/vrt-compare/**/*.ts"
     ],
     "references": [
         {
