+++ @
@@ -1,0 +1,30 @@
+chore(a11y-testing-spike): move implementation to 2nd gen, move docs
+
+# Please enter the commit message for your changes. Lines starting
+# with '#' will be ignored, and an empty message aborts the commit.
+#
+# Date:      Tue Nov 4 15:16:43 2025 -0800
+#
+# interactive rebase in progress; onto 4266ccd96e
+# Last commands done (2 commands done):
+#    pick c6e0dc9c05 # chore(a11y-testing-spike): move implementation to 2nd gen, move docs
+#    squash a664498355 # chore(a11y-testing-spike): move implementation to 2nd gen, move docs
+# No commands remaining.
+# You are currently rebasing branch 'cdransf/a11y-testing-spike' on '4266ccd96e'.
+#
+# Changes to be committed:
+#	modified:   1st-gen/package.json
+#	modified:   1st-gen/packages/badge/test/badge.a11y.spec.ts
+#	modified:   1st-gen/packages/status-light/test/status-light.a11y.spec.ts
+#	modified:   2nd-gen/package.json
+#	modified:   2nd-gen/packages/swc/components/badge/test/badge.a11y.spec.ts
+#	modified:   2nd-gen/packages/swc/components/status-light/test/status-light.a11y.spec.ts
+#	renamed:    1st-gen/playwright.a11y.config.ts -> 2nd-gen/playwright.a11y.config.ts
+#	renamed:    1st-gen/test/a11y-helpers.ts -> 2nd-gen/test/a11y-helpers.ts
+#	modified:   CONTRIBUTOR-DOCS/01_contributor-guides/03_working-in-the-swc-repo.md
+#	modified:   CONTRIBUTOR-DOCS/01_contributor-guides/06_releasing-swc.md
+#	renamed:    ACCESSIBILITY_TESTING.md -> CONTRIBUTOR-DOCS/01_contributor-guides/09_accessibility-testing.md
+#	modified:   CONTRIBUTOR-DOCS/01_contributor-guides/README.md
+#	modified:   CONTRIBUTOR-DOCS/README.md
+#	modified:   yarn.lock
+#
